Added boilerplate and fetch
The function I tested in Node, seems to be anough to work as frontend
simple search too. May be too slow for mobile devices, may not, that has
to be tested. The fetch function worked, now I have to make it read all
files listed in the json at the root of the book directory, and make the
code search through all files. Other problems to be solved are: Dealing
with files that can be either remote or local (sending it all to the
local storage might solve that), also implementing a scrollToLocation
function (allowing the search page to redirect the user to the part of
the book he wants), also saving the last search result.